requirements.txt
fastapi
uvicorn[standard]
pydantic
pydantic-settings
boto3
openai
twilio
python-multipart
sendgid
motor
amazon-transcribe-streaming